Generic test runner: Migrate Select (15/n)
[ghstack-poisoned]
