+++ packages/component-driver-html-test/__tests__/HTMLSelectDriver.e2e.test.ts
@@ -1,24 +1,7 @@
-import { createTestEngine } from '@atomic-testing/playwright';
-import { expect, test } from '@playwright/test';
+import { getTestRunnerInterface, playWrightTestFrameworkMapper } from '@atomic-testing/playwright';
+import { testRunner } from '@atomic-testing/test-runner';
+import { multipleSelectTestSuite, singleSelectTestSuite } from '../src/examples';
 
-import { multipleSelectExample, singleSelectExample } from '../src/examples/HTMLSelect.examples';
+testRunner(singleSelectTestSuite, playWrightTestFrameworkMapper, getTestRunnerInterface());
 
-test('Single Select', async ({ page }) => {
-  await page.goto('/select');
-  const testEngine = createTestEngine(page, singleSelectExample.scene);
-  const targetValue = '3';
-  await testEngine.parts.select.setValue(targetValue);
-  const val = await testEngine.parts.select.getValue();
-  expect(val).toBe(targetValue);
-  await testEngine.cleanUp();
-});
-
-test('Multiple Select', async ({ page }) => {
-  await page.goto('/select');
-  const testEngine = createTestEngine(page, multipleSelectExample.scene);
-  const targetValue = ['3', '5'];
-  await testEngine.parts.select.setValue(targetValue);
-  const val = await testEngine.parts.select.getValue();
-  expect(val).toEqual(targetValue);
-  await testEngine.cleanUp();
-});
+testRunner(multipleSelectTestSuite, playWrightTestFrameworkMapper, getTestRunnerInterface());
