--- packages/component-driver-html-test/src/examples/HTMLSelect.examples.tsx
@@ -1,66 +1,0 @@
-import { HTMLSelectDriver } from '@atomic-testing/component-driver-html';
-import { byName, IExampleUnit, ScenePart } from '@atomic-testing/core';
-import React from 'react';
-
-export const SingleSelectExample = () => {
-  return (
-    <React.Fragment>
-      <form>
-        <select name="single-select">
-          <option value="1">One</option>
-          <option value="2">Two</option>
-          <option value="3">Three</option>
-          <option value="4">Four</option>
-          <option value="5">Five</option>
-        </select>
-      </form>
-    </React.Fragment>
-  );
-};
-
-export const singleSelectExampleScenePart = {
-  select: {
-    locator: byName('single-select'),
-    driver: HTMLSelectDriver,
-  },
-} satisfies ScenePart;
-
-export const singleSelectExample: IExampleUnit<typeof singleSelectExampleScenePart, JSX.Element> = {
-  title: 'Single Select',
-  scene: singleSelectExampleScenePart,
-  ui: <SingleSelectExample />,
-};
-
-export const MultipleSelectExample = () => {
-  return (
-    <React.Fragment>
-      <form>
-        <select name="multiple-select" multiple size={5}>
-          <option value="1">One</option>
-          <option value="2">Two</option>
-          <option value="3">Three</option>
-          <option value="4">Four</option>
-          <option value="5">Five</option>
-        </select>
-      </form>
-    </React.Fragment>
-  );
-};
-
-export const multipleSelectExampleScenePart = {
-  select: {
-    locator: byName('multiple-select'),
-    driver: HTMLSelectDriver,
-  },
-} satisfies ScenePart;
-
-export const multipleSelectExample: IExampleUnit<typeof multipleSelectExampleScenePart, JSX.Element> = {
-  title: 'Multiple Select',
-  scene: multipleSelectExampleScenePart,
-  ui: <MultipleSelectExample />,
-};
-
-export const selectExamples = [singleSelectExample, multipleSelectExample] satisfies IExampleUnit<
-  ScenePart,
-  JSX.Element
->[];
