+++ packages/component-driver-html-test/src/examples/index.ts
@@ -1,4 +1,4 @@
 export * from './checkbox';
-export * from './HTMLSelect.examples';
 export * from './input';
 export * from './radioButtonGroup';
+export * from './select';
